Allow a receiver parameter on constructors
