+++ javaparser-core/src/main/java/com/github/javaparser/ast/body/MethodDeclaration.java
@@ -36,9 +36,7 @@ import com.github.javaparser.ast.type.Type;
 import com.github.javaparser.ast.type.TypeParameter;
 import com.github.javaparser.ast.visitor.GenericVisitor;
 import com.github.javaparser.ast.visitor.VoidVisitor;
-import java.util.Arrays;
 import java.util.EnumSet;
-import java.util.List;
 import java.util.Optional;
 import static com.github.javaparser.ast.Modifier.*;
 import static com.github.javaparser.utils.Utils.assertNotNull;
@@ -66,8 +64,6 @@ public final class MethodDeclaration extends CallableDeclaration<MethodDeclarati
     private Type type;
 
     private BlockStmt body;
-
-    private ReceiverParameter receiverParameter;
 
     public MethodDeclaration() {
         this(null, EnumSet.noneOf(Modifier.class), new NodeList<>(), new NodeList<>(), new ClassOrInterfaceType(), new SimpleName(), new NodeList<>(), new NodeList<>(), new BlockStmt(), null);
@@ -102,7 +98,7 @@ public final class MethodDeclaration extends CallableDeclaration<MethodDeclarati
      */
     @Generated("com.github.javaparser.generator.core.node.MainConstructorGenerator")
     public MethodDeclaration(TokenRange tokenRange, EnumSet<Modifier> modifiers, NodeList<AnnotationExpr> annotations, NodeList<TypeParameter> typeParameters, Type type, SimpleName name, NodeList<Parameter> parameters, NodeList<ReferenceType> thrownExceptions, BlockStmt body, ReceiverParameter receiverParameter) {
-        super(tokenRange, modifiers, annotations, typeParameters, name, parameters, thrownExceptions);
+        super(tokenRange, modifiers, annotations, typeParameters, name, parameters, thrownExceptions, receiverParameter);
         setType(type);
         setBody(body);
         setReceiverParameter(receiverParameter);
@@ -282,12 +278,6 @@ public final class MethodDeclaration extends CallableDeclaration<MethodDeclarati
                 return true;
             }
         }
-        if (receiverParameter != null) {
-            if (node == receiverParameter) {
-                removeReceiverParameter();
-                return true;
-            }
-        }
         return super.remove(node);
     }
 
@@ -319,12 +309,6 @@ public final class MethodDeclaration extends CallableDeclaration<MethodDeclarati
                 return true;
             }
         }
-        if (receiverParameter != null) {
-            if (node == receiverParameter) {
-                setReceiverParameter((ReceiverParameter) replacementNode);
-                return true;
-            }
-        }
         if (node == type) {
             setType((Type) replacementNode);
             return true;
@@ -352,28 +336,5 @@ public final class MethodDeclaration extends CallableDeclaration<MethodDeclarati
     @Override
     public ResolvedMethodDeclaration resolve() {
         return getSymbolResolver().resolveDeclaration(this, ResolvedMethodDeclaration.class);
-    }
-
-    @Generated("com.github.javaparser.generator.core.node.PropertyGenerator")
-    public Optional<ReceiverParameter> getReceiverParameter() {
-        return Optional.ofNullable(receiverParameter);
-    }
-
-    @Generated("com.github.javaparser.generator.core.node.PropertyGenerator")
-    public MethodDeclaration setReceiverParameter(final ReceiverParameter receiverParameter) {
-        if (receiverParameter == this.receiverParameter) {
-            return (MethodDeclaration) this;
-        }
-        notifyPropertyChange(ObservableProperty.RECEIVER_PARAMETER, this.receiverParameter, receiverParameter);
-        if (this.receiverParameter != null)
-            this.receiverParameter.setParentNode(null);
-        this.receiverParameter = receiverParameter;
-        setAsParentNodeOf(receiverParameter);
-        return this;
-    }
-
-    @Generated("com.github.javaparser.generator.core.node.RemoveMethodGenerator")
-    public MethodDeclaration removeReceiverParameter() {
-        return setReceiverParameter((ReceiverParameter) null);
     }
 }
