+++ javaparser-core/src/main/java/com/github/javaparser/metamodel/CallableDeclarationMetaModel.java
@@ -19,6 +19,8 @@ public class CallableDeclarationMetaModel extends BodyDeclarationMetaModel {
 
     public PropertyMetaModel parametersPropertyMetaModel;
 
+    public PropertyMetaModel receiverParameterPropertyMetaModel;
+
     public PropertyMetaModel thrownExceptionsPropertyMetaModel;
 
     public PropertyMetaModel typeParametersPropertyMetaModel;
